+++ scripts/build_env/appregdef_render.py
@@ -4,8 +4,8 @@ from pathlib import Path
 from envgenehelper import getenv_with_error, get_template_dirs, get_env_instances_dir, find_cloud_passport_definition, \
     update_generated_versions, NamespaceRole, BUILD_ENV_TAG
 
-from env_template.process_env_template import process_env_template
-from render_config_env import EnvGenerator
+from build_env.env_template.process_env_template import process_env_template
+from build_env.render_config_env import EnvGenerator
 
 
 def run_appregdef_render():
