+++ scripts/tests/sd/test_process_sd_artifact.py
@@ -7,7 +7,7 @@ from envgenehelper.env_helper import Environment
 from envgenehelper.models import OperationType
 
 from tests.base_test import BaseTest
-from test_sd_helpers import do_prerequisites, assert_sd_contents, load_test_pipeline_sd_data
+from tests.sd.test_sd_helpers import do_prerequisites, assert_sd_contents, load_test_pipeline_sd_data
 
 os.environ['ENVIRONMENT_NAME'] = "temporary"
 os.environ['CLUSTER_NAME'] = "temporary"
